project sih/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
mysql-connector-python==8.2.0
Werkzeug==2.3.7
